+++ R/theme_gtsummary_roche.R
@@ -5,6 +5,8 @@
 #' - Calls the [`gtsummary::theme_gtsummary_compact()`] theme.
 #' - Uses `label_roche_pvalue()` as the default formatting function for all p-values.
 #' - Defaults to a mono-spaced font for gt tables.
+#' - Font size defaults are 8 points for all the table by the footers that are 7 points.
+#' - Border defaults to `flextable::fp_border_default(width = 0.5)`.
 #'
 #' @inheritParams gtsummary::theme_gtsummary_compact
 #'
@@ -17,10 +19,31 @@
 #' tbl_demographics(trial, by = trt, include = c(age, grade))
 #'
 #' reset_gtsummary_theme()
-theme_gtsummary_roche <- function(set_theme = TRUE, font_size = NULL) {
-  # start with the compact theme -----------------------------------------------
-  lst_theme <- gtsummary::theme_gtsummary_compact(set_theme = FALSE, font_size = font_size)
-  lst_theme$`pkgwide-str:theme_name` <- "Roche"
+theme_gtsummary_roche <- function(set_theme = TRUE) {
+  # {crane} defaults
+  font_size <- 8
+  font_size_gt <- 13 # gt counts size points differently
+  border <- flextable::fp_border_default(width = 0.5)
+
+  # Initialization with compact gt options -------------------------------------
+  lst_theme <-
+    list(
+      "pkgwide-str:theme_name" = "Roche",
+      # compact gt tables
+      "as_gt-lst:addl_cmds" = list(
+        tab_spanner = rlang::expr(
+          gt::tab_options(
+            table.font.size = !!font_size_gt,
+            data_row.padding = gt::px(1),
+            summary_row.padding = gt::px(1),
+            grand_summary_row.padding = gt::px(1),
+            footnotes.padding = gt::px(1),
+            source_notes.padding = gt::px(1),
+            row_group.padding = gt::px(1)
+          )
+        )
+      )
+    )
 
   # updating with some pharma-specific bits ------------------------------------
   lst_theme <- lst_theme |>
@@ -32,21 +55,53 @@ theme_gtsummary_roche <- function(set_theme = TRUE, font_size = NULL) {
       )
     )
 
-  # use arial font for flextables ----------------------------------------------
+  # {flextable} options --------------------------------------------------------
   lst_theme$`as_flex_table-lst:addl_cmds` <-
     c(
       lst_theme$`as_flex_table-lst:addl_cmds`,
-      list(valign = rlang::expr(flextable::font(fontname = "Arial", part = "all")))
+      list(
+        fontsize =
+          list(
+            rlang::expr(flextable::fontsize(size = !!font_size, part = "all")),
+            rlang::expr(flextable::fontsize(size = !!(font_size - 1), part = "footer"))
+          ),
+        border = list(
+          rlang::expr(flextable::border_outer(part = "body", border = !!border)),
+          rlang::expr(flextable::border_outer(part = "header", border = !!border))
+        ),
+        valign = list( # valign only because it will append to to last commands
+          rlang::expr(flextable::valign(valign = "top", part = "all")),
+          rlang::expr(flextable::font(fontname = "Arial", part = "all")),
+          rlang::expr(flextable::padding(padding.top = 0, part = "all")),
+          rlang::expr(flextable::padding(padding.bottom = 0, part = "all")),
+          rlang::expr(flextable::line_spacing(space = 1, part = "all")),
+          rlang::expr(flextable::set_table_properties(layout = "autofit"))
+        )
+      )
     )
-
 
   # add a monospace font for gt ------------------------------------------------
   lst_theme$`as_gt-lst:addl_cmds` <-
     c(
       lst_theme$`as_gt-lst:addl_cmds`,
-      list(tab_spanner = rlang::expr(gt::opt_table_font(stack = "monospace-code")))
+      list(
+        cols_hide =
+          list(
+            rlang::expr(gt::opt_table_font(font = "arial")),
+            rlang::expr(
+              gt::tab_options(
+                table.font.size = 13,
+                data_row.padding = gt::px(1),
+                summary_row.padding = gt::px(1),
+                grand_summary_row.padding = gt::px(1),
+                footnotes.padding = gt::px(1),
+                source_notes.padding = gt::px(1),
+                row_group.padding = gt::px(1)
+              )
+            )
+          )
+      )
     )
-
 
   # finishing up ---------------------------------------------------------------
   if (set_theme == TRUE) gtsummary::set_gtsummary_theme(lst_theme)